requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==4.0.0
gunicorn==21.2.0
ultralytics==8.0.43
opencv-python-headless==4.8.1.78
torch==2.0.1
torchvision==0.15.2
python-dotenv==1.0.0
pyyaml==6.0.1
pillow==10.0.1
numpy==1.24.4
matplotlib==3.7.2
pandas==2.0.3
scipy==1.11.3
jinja2==3.1.2
